Split data.ts into separate files
